refactor: 컴포넌트 응집성을 높이기 위한 폴더구조 변경
