+++ src/components/@common/input.js
@@ -1,13 +1,20 @@
 import { $ } from "../../utils/domHelpers";
 
 const input = (props) => {
-  const { labelText, id, isRequired = false, spanText = "" } = props;
+  const {
+    labelText,
+    id,
+    isRequired = false,
+    spanText = "",
+    onChange,
+    onInput,
+  } = props;
 
   const required = isRequired ? "required" : "";
 
   return `
         <label for="${id} text-caption">${labelText}</label>
-        <input type="text" name=${id} id=${id} ${required} maxLength='20'/>
+        <input type="text" name=${id} id=${id} ${required} maxLength='20' onchange=${onChange} oninput=${onInput}/>
         <span class="help-text text-caption">${spanText}</span>
     `;
 };
