+++ src/views/mainPage/mainPage.js
@@ -1,12 +1,12 @@
 import errorContainer from "../../components/errorContainer.js";
 import fallbackComponent from "../../components/fallbackComponent.js";
-import buttonContainer from "./components/buttonContainer";
-import categorySelect from "./components/categorySelect";
-import description from "./components/description";
-import distanceSelect from "./components/distanceSelect";
-import linkInput from "./components/linkInput";
-import nameInput from "./components/nameInput";
-import restaurantList from "./components/restaurantList";
+import buttonContainer from "../../components/buttonContainer/buttonContainer.js";
+import categorySelect from "../../components/categorySelect/categorySelect.js";
+import description from "../../components/description/description.js";
+import distanceSelect from "../../components/distanceSelect/distanceSelect.js";
+import linkInput from "../../components/linkInput/linkInput.js";
+import nameInput from "../../components/nameInput/nameInput.js";
+import restaurantList from "../../components/restaurantList/restaurantList.js";
 
 const renderMainPage = () => {
   errorContainer(() => restaurantList(), fallbackComponent);
